Font embedding tries to replace type 3 fonts, if PS file is available.

+++ embed-pdf-fonts.sh
@@ -11,14 +11,45 @@
 filename=$GEDIT_CURRENT_DOCUMENT_NAME
 basename=${filename%.*}
 
-# Embedd fonts into a temp PDF file
-ps2pdf -dPDFSETTINGS=/prepress "$basename.pdf" temp-embedding.pdf
+if [ -f "$basename.ps" ];
+then
 
-# Remove original PDF
-rm -f "$basename.pdf"
+  #Convert PDF to PS
+  #pdftops -origpagesizes "$basename.pdf" "$basename.ps"
 
-# Rename temp PDF
-mv -f temp-embedding.pdf "$basename.pdf"
+  # Remove Type 3 fonts
+  pkfix "$basename.ps" "$basename-fix.ps"
+
+  # Embedd fonts into a temp PDF file
+  ps2pdf -dPDFSETTINGS=/prepress \
+         -dCompatibility=1.2 \
+         "$basename-fix.ps" \
+         "$basename.pdf"
+
+# Remove original and fixed PS files
+rm -f "$basename.ps" "$basename-fix.ps"
+
+fi
+
+if [ ! -f "$basename.ps" ] && [ -f "$basename.pdf" ];
+then
+
+  ## Rename temp PDF
+  #mv -f temp-embedding.pdf "$basename.pdf"
+
+  # Embedd fonts into a temp PDF file
+  ps2pdf -dPDFSETTINGS=/prepress \
+         -dCompatibility=1.3 \
+         "$basename.pdf" \
+         "$basename-embed.pdf"
+
+  # Remove original PDF
+  rm -f "$basename.pdf"
+
+  # Rename temp PDF
+  mv -f "$basename-embed.pdf" "$basename.pdf"
+
+fi
 
 # Open PDF
 xdg-open "$basename.pdf" &
